Disabled options across multiple instances
The script tends to include all commands that are included via a js
file.  However in the case of multiple instances it may be that some
commands are enabled for one instance and commands are disabled for
another instance.  This little bit of code will ensure any disabled
options don't show up even when an include file is present.