Update elFinder.js
Fix for issue #1146.
The connector sends to the client the value of upload_max_filesize in the php.ini. If it is set to 0, the uplMaxSize variable is set to "0" (as string, not numeric) and when getting the BYTES_PER_CHUNK, the or between "0" and a number, it returns "0". So 0 minus 8190 is negative and thus it enters an infinite loop.