[php:core:security] disallow local loopback on URL upload and add an option `urlUploadFilter`
At the time of uploading the URL, HTTP content acquisition to the local
loopback address is not preferable for security, so we will not allow it
and we have added a new option `urlUploadFilter`.

```php
$opts = array(
    'urlUploadFilter' => function($url, $elFinderInstance) {
    	// It checks $url and returns true if it is acceptable, and
    	// false if it declines it.
    	$res = preg_match('~^https?://example.com~i', $url);
    	return $res;
    },
    'roots'  => array(
        array(
            'driver' => 'LocalFileSystem',
            'path'   => '/path/to/files/',
            'URL'    => 'http://localhost/to/files/'
        )
    )
);
// run elFinder
$connector = new elFinderConnector(new elFinder($opts));
$connector->run();
```

This commitment was received from Do Ha Anh of Viettel Cyber Security
Center and I got an idea from him.

Again, I would like to express my special thanks to Do Ha Anh of Viettel
Cyber Security Center.