Fixes #3689, fm.sync removes unavailable volumes. (#3690)
* Fixes Studio-42/elFinder#3689, fm.sync removes unavailable volumes.

When a request is made to the server with the open command and tree=1,
the server must reply with a list of all volumes. If the server's
response does not include a volume ElFinder already knows about, that
volume should be treated as not available anymore and get removed
from the file explorer UI.

* Fixes Studio-42/elFinder#3689, fm.sync removes unavailable volumes.

This is an amendment to the previous commit. When invoking the refresh
action on a non-root folder, ElFinder also makes a "tree" request to the
server and adds all root volumes it currently knows about. Sinc other
parts of the code might rely upon this behavior (such as tree.js), remove
non-existent volumes from the response to the tree command in the fm.sync
method.

---------

Co-authored-by: Andre Wachsmuth <awa@xima.de>