Fix build for Windows environment (#3699)
mkdir command in Windows does not have a "-p" option. Creating
the parents seems unnecessary, as the parent folder is the project
root and that should always exist. So simply remove the "-p" option.

Windows uses backslashes as the path separator. Replace backslash
with forward slash to normalize the paths.

Co-authored-by: Andre Wachsmuth <awa@xima.de>